Encapsulate and move cruft out of ogc.py for readability and testability.
